requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
beautifulsoup4==4.13.4
lxml==5.4.0
python-dotenv==1.1.1
numpy==2.2.6
sentence-transformers==3.4.1
huggingface-hub>=0.34.0,<1.0
faiss-cpu==1.10.0
pydantic==2.11.7
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.14.0
snowballstemmer==2.2.0
python-multipart==0.0.20
pypdf==5.4.0
python-docx==1.1.2
